Fix s3 pause/continue chunked multiple bug.
Reproduce:
s3.XhrUploadHandler
options:
, chunking: { enabled: true, concurrent: { enabled: true } }
, maxConnections: 3
, multiple: true
, resume: { enabled: true}

Chrome: Version 56.0.2924.87 (64-bit)

Begin uploading a large file. Pause and continue the upload a few times
and notice that immediately or eventually (it's sporadic) the upload will
be paused but chunks will still upload in the background.

Solution: the paused state is not cleared on upload continue. initHeaders
promise success seems to be called even though the XHR is aborted. Added
file status pause check here to prevent an XHR PUT when file hsa been
paused.